+++ Collector/ProfileClientFactory.php
@@ -6,6 +6,7 @@ use Http\Client\Common\FlexibleHttpClient;
 use Http\Client\HttpAsyncClient;
 use Http\Client\HttpClient;
 use Http\HttplugBundle\ClientFactory\ClientFactory;
+use Symfony\Component\Stopwatch\Stopwatch;
 
 /**
  * The ProfileClientFactory decorates any ClientFactory and returns the created client decorated by a ProfileClient.
@@ -32,11 +33,17 @@ class ProfileClientFactory implements ClientFactory
     private $formatter;
 
     /**
+     * @var Stopwatch
+     */
+    private $stopwatch;
+
+    /**
      * @param ClientFactory|callable $factory
      * @param Collector              $collector
      * @param Formatter              $formatter
+     * @param Stopwatch              $stopwatch
      */
-    public function __construct($factory, Collector $collector, Formatter $formatter)
+    public function __construct($factory, Collector $collector, Formatter $formatter, Stopwatch $stopwatch)
     {
         if (!$factory instanceof ClientFactory && !is_callable($factory)) {
             throw new \RuntimeException(sprintf('First argument to ProfileClientFactory::__construct must be a "%s" or a callable.', ClientFactory::class));
@@ -44,6 +51,7 @@ class ProfileClientFactory implements ClientFactory
         $this->factory = $factory;
         $this->collector = $collector;
         $this->formatter = $formatter;
+        $this->stopwatch = $stopwatch;
     }
 
     /**
@@ -57,6 +65,6 @@ class ProfileClientFactory implements ClientFactory
             $client = new FlexibleHttpClient($client);
         }
 
-        return new ProfileClient($client, $this->collector, $this->formatter);
+        return new ProfileClient($client, $this->collector, $this->formatter, $this->stopwatch);
     }
 }