+++ DependencyInjection/HttplugExtension.php
@@ -279,11 +279,6 @@ class HttplugExtension extends Extension
 
         $pluginClientOptions = [];
         if ($profiling) {
-            // Add the stopwatch plugin
-            if (!in_array('httplug.plugin.stopwatch', $arguments['plugins'])) {
-                array_unshift($plugins, 'httplug.plugin.stopwatch');
-            }
-
             //Decorate each plugin with a ProfilePlugin instance.
             foreach ($plugins as $pluginServiceId) {
                 $this->decoratePluginWithProfilePlugin($container, $pluginServiceId);
@@ -561,6 +556,7 @@ class HttplugExtension extends Extension
                     $factory,
                     new Reference('httplug.collector.collector'),
                     new Reference('httplug.collector.formatter'),
+                    new Reference('debug.stopwatch'),
                 ]);
             $factory = new Reference($factoryServiceId);
         }