+++ Tests/Unit/DependencyInjection/HttplugExtensionTest.php
@@ -117,7 +117,6 @@ class HttplugExtensionTest extends AbstractExtensionTestCase
 
         $plugins = [
             'httplug.client.acme.plugin.stack',
-            'httplug.plugin.stopwatch',
             'httplug.client.acme.plugin.decoder',
             'httplug.plugin.redirect',
             'httplug.client.acme.plugin.add_host',